requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
flatbuffers==23.3.3
fonttools==4.39.4
gast==0.4.0
GDAL==3.6.4
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.51.3
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
idna==3.4
jax==0.4.6
joblib==1.2.0
keras==2.12.0
keras-rl==0.4.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mysql-connector-python==8.0.32
neat-python==0.92
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.8.10
packaging==23.0
Pillow==9.5.0
poke-env==0.6.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2023.5.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0rc1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorrec==0.25.4
termcolor==2.2.0
textblob==0.17.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
websockets==11.0.1
Werkzeug==2.2.3
wrapt==1.14.1
